+++ src/main/java/kr/hhplus/be/server/interfaces/api/option/dto/OptionDto.java
@@ -1,6 +1,8 @@
 package kr.hhplus.be.server.interfaces.api.option.dto;
 
+import java.math.BigDecimal;
+
 public record OptionDto(
         String name,
-        int additionalCost
+        BigDecimal additionalCost
 ) { }
