+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/dto/OrderProductResponse.java
@@ -2,7 +2,9 @@ package kr.hhplus.be.server.interfaces.api.order.dto;
 
 import kr.hhplus.be.server.domain.order.OrderProduct;
 
-public record OrderProductResponse(Long productId, int quantity, int unitPoint) {
+import java.math.BigDecimal;
+
+public record OrderProductResponse(Long productId, int quantity, BigDecimal unitPoint) {
 
     public static OrderProductResponse from(OrderProduct op) {
         return new OrderProductResponse(op.getProductId(), op.getQuantity(), op.getUnitPoint());
