+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/dto/ChargePointRequestDto.java
@@ -1,7 +1,9 @@
 package kr.hhplus.be.server.interfaces.api.point.dto;
 
+import java.math.BigDecimal;
+
 public record ChargePointRequestDto(
         long userId,
-        int amount
+        BigDecimal amount
 ) {
 }
