+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/dto/PointHistoryResponseDto.java
@@ -1,8 +1,10 @@
 package kr.hhplus.be.server.interfaces.api.point.dto;
 
+import java.math.BigDecimal;
+
 public record PointHistoryResponseDto(
         Long id,
         Long userId,
-        int changeAmount,
+        BigDecimal changeAmount,
         String transactionType
 ) { }
