+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/dto/ProductDto.java
@@ -1,10 +1,12 @@
 package kr.hhplus.be.server.interfaces.api.product.dto;
 
 
+import java.math.BigDecimal;
+
 public record ProductDto(
         Long id,
         String itemName,
         String optionName,
-        double finalPrice
+        BigDecimal finalPrice
 ) {
 }
